Rename Shop Dto namespace to Dtos
