+++ src/WebApiLearning.Application/Shops/Commands/Update/UpdateShopCommand.cs
@@ -1,7 +1,7 @@
 using System.ComponentModel.DataAnnotations;
 using LanguageExt.Common;
 using MediatR;
-using WebApiLearning.Application.Shops.Dto;
+using WebApiLearning.Application.Shops.Dtos;
 
 namespace WebApiLearning.Application.Shops.Commands.Update;
