+++ src/WebApiLearning.Application/Shops/Commands/Update/UpdateShopRequestHandler.cs
@@ -1,6 +1,6 @@
 using LanguageExt.Common;
 using MediatR;
-using WebApiLearning.Application.Shops.Dto;
+using WebApiLearning.Application.Shops.Dtos;
 using WebApiLearning.Domain.Entities;
 using WebApiLearning.Domain.Repository;
