+++ src/WebApiLearning.Application/Shops/Dtos/ShopDto.cs
@@ -1,3 +1,3 @@
-﻿namespace WebApiLearning.Application.Shops.Dto;
+﻿namespace WebApiLearning.Application.Shops.Dtos;
 
 public record ShopDto(Guid Id, string Name, string Location);
