+++ src/WebApiLearning.Application/Shops/Queries/List/ListShopsHandler.cs
@@ -1,6 +1,6 @@
 using LanguageExt.Common;
 using MediatR;
-using WebApiLearning.Application.Shops.Dto;
+using WebApiLearning.Application.Shops.Dtos;
 using WebApiLearning.Domain.Repository;
 
 namespace WebApiLearning.Application.Shops.Queries.List;
