+++ src/WebApiLearning.Application/Shops/Queries/List/ListShopsQuery.cs
@@ -1,6 +1,6 @@
 using LanguageExt.Common;
 using MediatR;
-using WebApiLearning.Application.Shops.Dto;
+using WebApiLearning.Application.Shops.Dtos;
 
 namespace WebApiLearning.Application.Shops.Queries.List;
