+++ src/WebApiLearning.Application/WebApiLearning.Application.csproj
@@ -12,6 +12,7 @@
     <ProjectReference Include="..\WebApiLearning.Domain\WebApiLearning.Domain.csproj" />
   </ItemGroup>
   <ItemGroup>
+    <Folder Include="Inventories\Dtos\" />
     <Folder Include="Inventories\Queries\" />
   </ItemGroup>
 </Project>
